// frontend/src/app/core/api/model/transactionRequest.ts
/**
 * Le Tacos Backend API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface TransactionRequestDto { 
    type: TransactionRequestDto.TypeEnum;
    amount: number;
    transactionDate: string;
    categoryId: string;
    paymentMethod: TransactionRequestDto.PaymentMethodEnum;
    notes?: string;
    employeeId?: string;
    menuItemId?: string;
}
export namespace TransactionRequestDto {
    export const TypeEnum = {
        Income: 'INCOME',
        Expense: 'EXPENSE'
    } as const;
    export type TypeEnum = typeof TypeEnum[keyof typeof TypeEnum];
    export const PaymentMethodEnum = {
        Cash: 'CASH',
        Card: 'CARD',
        Wolt: 'WOLT',
        BankTransfer: 'BANK_TRANSFER',
        Other: 'OTHER'
    } as const;
    export type PaymentMethodEnum = typeof PaymentMethodEnum[keyof typeof PaymentMethodEnum];
}