// frontend/src/app/core/api/model/transactionCategoryResponse.ts
/**
 * Le Tacos Backend API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface TransactionCategoryResponseDto { 
    id?: string;
    type?: TransactionCategoryResponseDto.TypeEnum;
    name?: string;
    systemKey?: string;
    active?: boolean;
    displayOrder?: number;
}
export namespace TransactionCategoryResponseDto {
    export const TypeEnum = {
        Income: 'INCOME',
        Expense: 'EXPENSE'
    } as const;
    export type TypeEnum = typeof TypeEnum[keyof typeof TypeEnum];
}